dom/media/test/test_av1_high_professional_profiles.html

<!DOCTYPE html>
<html>
<head>
<title>Test High and Professional profile AV1 video files</title>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
<script type="application/javascript">

/** This test checks if 10- and 12-bit AV1 video plays. */
const av1Files = ["av1-yuv444p10.mp4", "av1-yuv422p10.mp4", "av1-yuv444p12.mp4"];
add_task(async function testPlayAll() {
  await Promise.all(av1Files.map(
    (f) => {
      let v = document.createElement("video");
      document.body.appendChild(v);
      v.src = f;
      return v;
    }).map(
      (v) => v.play())).then(
        () => ok(true),
        () => ok(false));
});

</script>
</head>
<body>
</body>
</html>
